refactor: Extract DOM methods from index
